+++ entrypoints/popup/views/card/CardView.tsx
@@ -2,12 +2,12 @@ import { useState } from 'react';
 import { Button, Input, Label, TextField } from 'react-aria-components';
 import { FaArrowUpRightFromSquare, FaCirclePause, FaMagnifyingGlass, FaPlay, FaTrash, FaXmark } from 'react-icons/fa6';
 import { State as FsrsState } from 'ts-fsrs';
+import { bounceButton } from '@/entrypoints/popup/styles';
 import { useCardsQuery, usePauseCardMutation, useRemoveCardMutation } from '@/hooks/queries/cards';
 import { useTimedConfirmation } from '@/hooks/useTimedConfirmation';
 import type { Card } from '@/shared/cards';
 import type { Translations } from '@/shared/i18n';
 import { getLeetcodeProblemUrl } from '@/shared/leetcode';
-import { bounceButton } from '@/shared/styles';
 import { StreakCounter } from '../../components/StreakCounter';
 import { ViewLayout } from '../../components/ViewLayout';
 import { useI18n } from '../../contexts/I18nContext';
