+++ entrypoints/popup/views/card/components/CardNotes.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react';
 import { Button, Label, TextArea, TextField } from 'react-aria-components';
+import { bounceButton } from '@/entrypoints/popup/styles';
 import { useNoteEditor } from '@/hooks/useNoteEditor';
 import { NOTES_MAX_LENGTH } from '@/shared/notes';
-import { bounceButton } from '@/shared/styles';
 import { useI18n } from '../../../contexts/I18nContext';
 
 const MAX_TEXTAREA_HEIGHT = 160; // px, matches max-h-40
