+++ entrypoints/popup/views/home/NotesSection.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Button, Label, TextArea, TextField } from 'react-aria-components';
+import { bounceButton } from '@/entrypoints/popup/styles';
 import { useNoteEditor } from '@/hooks/useNoteEditor';
 import { NOTES_MAX_LENGTH } from '@/shared/notes';
-import { bounceButton } from '@/shared/styles';
 import { useI18n } from '../../contexts/I18nContext';
 
 interface NotesSectionProps {
